Rename MonitoringNetworkCapability to NetworkStatisticsCapability
