+++ extensions/bundles/ofnetwork/src/main/java/org/opennaas/extensions/ofnetwork/capability/statistics/INetworkStatisticsCapability.java
@@ -1,4 +1,4 @@
-package org.opennaas.extensions.ofnetwork.capability.monitoring;
+package org.opennaas.extensions.ofnetwork.capability.statistics;
 
 /*
  * #%L
@@ -20,8 +20,13 @@ package org.opennaas.extensions.ofnetwork.capability.monitoring;
  * #L%
  */
 
-import org.opennaas.core.resources.action.ActionResponse;
-import org.opennaas.core.resources.action.IActionSetDefinition;
+import javax.ws.rs.GET;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
+import javax.ws.rs.core.MediaType;
+
+import org.opennaas.core.resources.capability.CapabilityException;
+import org.opennaas.core.resources.capability.ICapability;
 import org.opennaas.extensions.ofnetwork.model.NetworkStatistics;
 
 /**
@@ -29,11 +34,12 @@ import org.opennaas.extensions.ofnetwork.model.NetworkStatistics;
  * @author Adrian Rosello Rey (i2CAT)
  * 
  */
-public class MonitoringNetworkActionSet implements IActionSetDefinition {
+@Path("/")
+public interface INetworkStatisticsCapability extends ICapability {
 
-	/**
-	 * An action that retrieves the statistics of all port of the network switches. It receives no parameters, and it returns a
-	 * {@link NetworkStatistics} object in the {@link ActionResponse#getResult()} method.
-	 */
-	public static final String	GET_NETWORK_STATISTICS	= "getNetworkStatistics";
+	@GET
+	@Path("readNetworkStatistics")
+	@Produces(MediaType.APPLICATION_XML)
+	public NetworkStatistics getNetworkStatistics() throws CapabilityException;
+
 }
