+++ extensions/bundles/ofnetwork/src/main/resources/OSGI-INF/blueprint/core.xml
@@ -47,14 +47,14 @@
 		</service-properties>
 	</service>
 	
-	<bean id="monitoringNetworkCapabilityFactory" class="org.opennaas.extensions.ofnetwork.capability.monitoring.MonitoringNetworkCapabilityFactory">
-		<property name="type" value="netmonitoring"/>
+	<bean id="networkStatisticsCapabilityFactory" class="org.opennaas.extensions.ofnetwork.capability.statistics.NetworkStatisticsCapabilityFactory">
+		<property name="type" value="netstatistics"/>
 	</bean>	
 	<!-- Register the Service as an OSGi Service -->
-	<service ref="monitoringNetworkCapabilityFactory" interface="org.opennaas.core.resources.capability.ICapabilityFactory">
+	<service ref="networkStatisticsCapabilityFactory" interface="org.opennaas.core.resources.capability.ICapabilityFactory">
 		<service-properties>
 			<!--  TODO it is necessary to use two parameters to get capability -->
-			<entry key="capability" value="netmonitoring"/>
+			<entry key="capability" value="netstatistics"/>
 			<entry key="capability.model" value="ofnetwork"/>
 			<entry key="capability.version" value="1.0.0"/>
 		</service-properties>
